requirements.txt
# Core data processing
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
python-dotenv>=0.19.0

# Financial data
yfinance>=0.2.0
fredapi>=0.5.0

# Machine Learning (free)
scikit-learn>=1.1.0
transformers>=4.21.0
torch>=1.12.0
sentence-transformers>=2.2.0

# NLP (free)
nltk>=3.7
textblob>=0.17.0
vaderSentiment>=3.3.0

# Dashboard
streamlit>=1.28.0
plotly>=5.11.0
altair>=4.2.0

# Utilities
feedparser>=6.0.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
schedule>=1.2.0

# Testing
pytest>=7.0.0
